+++ src/Plugin/Field/FieldType/TemplateWhispererFieldItem.php
@@ -147,7 +147,7 @@ class TemplateWhispererFieldItem extends FieldItemBase {
       $original_ids = [];
       $langcode = $this->getLangcode();
       $original = $entity->original;
-      if ($original->hasTranslation($langcode)) {
+      if ($original && $original->hasTranslation($langcode)) {
         $original_items = $original->getTranslation($langcode)->{$field_name};
         foreach ($original_items as $item) {
           $original_ids[] = $item->target_id;