+++ src/TwigExtension/TwigExtension.php
@@ -41,7 +41,7 @@ class TwigExtension extends AbstractExtension {
   /**
    * {@inheritdoc}
    */
-  public function getFunctions() {
+  public function getFunctions(): array {
     return [
       new TwigFunction('tw_suggestion_entities', $this->getEntitiesFromSuggestion(...)),
     ];